+++ androidApp/src/main/java/com/gorosoft/bookme/now/android/MainActivity.kt
@@ -12,7 +12,7 @@ import com.gorosoft.bookme.now.android.ui.account_setup.LoginScreen
 import com.gorosoft.bookme.now.android.ui.account_setup.create_your_profile.CreateYourProfileScreen
 import com.gorosoft.bookme.now.android.ui.account_setup.enter_otp.EnterOtpScreen
 import com.gorosoft.bookme.now.android.ui.main_screen.MainScreen
-import com.gorosoft.bookme.now.android.ui.main_screen.profile.EditProfileScreen
+import com.gorosoft.bookme.now.android.ui.main_screen.profile.edit_profile.EditProfileScreen
 import com.gorosoft.bookme.now.android.ui.main_screen.profile.choose_language.ChooseLanguageScreen
 import com.gorosoft.bookme.now.android.ui.onboarding.TutorialCarouselScreen
 import com.gorosoft.bookme.now.android.ui.onboarding.WelcomeScreen
