+++ androidApp/src/main/java/com/gorosoft/bookme/now/android/ui/main_screen/profile/ProfileScreen.kt
@@ -40,6 +40,8 @@ fun ProfileScreen(
     ProfileScreenContent(
         navigateToLanguageScreen = {
             navController.navigate(NavGraphDestination.ChooseLanguage.route)
+        },
+        navigateToEditProfileScreen = {
             navController.navigate(NavGraphDestination.EditProfile.route)
         }
     )
