+++ androidApp/src/main/java/com/gorosoft/bookme/now/android/ui/main_screen/profile/edit_profile/EditProfileScreenViewModel.java
@@ -1,0 +1,4 @@
+package com.gorosoft.bookme.now.android.ui.main_screen.profile.edit_profile;
+
+public class EditProfileScreenViewModel {
+}
